backend-service/requirements.txt
blinker==1.6.2
click==8.1.3
demo-nodes-py==0.9.4
domain-coordinator==0.9.2
Flask==2.3.2
Flask-Cors==3.0.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
osrf-pycommon==0.1.11
RPi.GPIO==0.7.1
pysqlite3==0.5.1
pydub==0.25.1
pygame==1.9.6